MLOps Based EfficientNetV2
The files named efficient_Net_V2_MLOps.py, serve.py and train.py needs to be opened in the IDE simultaneously. The user change the file path and model saving path in the the code. To train the algorithm, user should run train.py. The code has mlflow tracking functionality and the metrics can be saved and extracted from the mlflow UI on the local machine. Serve.py is used for running the endpoint on a local server in the web brower